Second Commit : Exception Handling On Server Side

+++ UserNotesSpringBootApp/UserNotesSpringBootApp/src/main/java/com/neeraj/UserNotesSpringBootApp/controller/NotesCategoryController.java
@@ -2,6 +2,8 @@ package com.neeraj.UserNotesSpringBootApp.controller;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.neeraj.UserNotesSpringBootApp.Exception.CategoryIdNotFoundException;
 import com.neeraj.UserNotesSpringBootApp.Service.NotesCategoryDAO;
 import com.neeraj.UserNotesSpringBootApp.Service.NotesDAO;
 import com.neeraj.UserNotesSpringBootApp.model.Notes;
@@ -33,49 +36,66 @@ public class NotesCategoryController
 	
 	//get all Category List
 	@GetMapping("/categories")
-	public List<NotesCategory> getAll()
+	public List<NotesCategory> getAll() throws Exception
 	{
+		try
+		{
 		return notesCategoryDAO.getAllCategory();
-	}
+		}
+		catch(Exception e)
+		{
+			throw new Exception();
+		}
+		
+		}
 	
 	
 	// Add new Category
 	@PostMapping("/categories/")
-	public ResponseEntity saveCategory(@RequestBody NotesCategory notesCategory)
+	public ResponseEntity<NotesCategory> saveCategory(@Valid @RequestBody NotesCategory notesCategory)
 	{
+		NotesCategory nCategory=notesCategoryDAO.saveCategory(notesCategory);
 		
-		return new ResponseEntity(notesCategoryDAO.saveCategory(notesCategory),HttpStatus.OK);
+		return new ResponseEntity<NotesCategory>(nCategory,HttpStatus.OK);
 	}
 	
 	
 	//Edit Category
 	@PutMapping("/categories/{id}")
-	public ResponseEntity updateCategory(@PathVariable(value="id") int categoryId ,@RequestBody NotesCategory notesCategory)
+	public ResponseEntity<NotesCategory> updateCategory(@Valid @PathVariable(value="id") int categoryId ,@Valid @RequestBody NotesCategory notesCategory) throws CategoryIdNotFoundException
 	{
 		NotesCategory notes=notesCategoryDAO.findCategory(categoryId);
 		
 		if(notes==null)
 		{
-			ResponseEntity.notFound().build();
+			throw new CategoryIdNotFoundException("Category Id :"+categoryId+" Not Found");
 		}
 		
 		notes.setCategoryName(notesCategory.getCategoryName());
 		
-		return new ResponseEntity(notesCategoryDAO.saveCategory(notes),HttpStatus.OK);
+		return new ResponseEntity<NotesCategory>(notesCategoryDAO.saveCategory(notes),HttpStatus.OK);
 	}
 	
 	
 	//Delete Category
 	@DeleteMapping("/categories/{id}")
-	public void deteleCategory(@PathVariable(value="id") int category_id)
+	public void deteleCategory(@Valid @PathVariable(value="id") int category_id) throws CategoryIdNotFoundException
 	{
+	
+       NotesCategory notesCategory=notesCategoryDAO.findCategory(category_id);
+		
+		if(notesCategory==null)
+		{
+			throw new CategoryIdNotFoundException("Category Id :"+category_id+" Not Found");
+		}
+		
+		
+		//find all the Note which belongs to this particular CategoryID 
 		List<Notes> note=notesDao.FindByCategoryID(category_id);
 		notesDao.deleteByCategoryId(note);
 		
-		NotesCategory notesCategory=notesCategoryDAO.findCategory(category_id);
+		//delete Category
 		notesCategoryDAO.deleteCategory(notesCategory);
-		
-		
-		
+			
 	}
 }
